+++ src/app_version/index.d.ts
@@ -1,6 +1,6 @@
 import type { be_computed_pair_T } from '@ctx-core/nanostores'
 export declare const [
-	app_version__,
+	app_version$_,
 	app_version_,
 ]:be_computed_pair_T<string>
-export { app_version__ as app_version$_ }
+export { app_version$_ as app_version__ }