+++ src/app_version/index.js
@@ -1,7 +1,7 @@
 import { CACHE_VERSION__, VERSION__ } from '@ctx-core/env'
 import { be_computed_pair_, computed_ } from '@ctx-core/nanostores'
 export const [
-	app_version__,
+	app_version$_,
 	app_version_,
 ] = be_computed_pair_('app_version_', ctx=>
 	computed_([
@@ -12,4 +12,4 @@ export const [
 			CACHE_VERSION
 			|| VERSION
 			|| Math.random().toString()))
-export { app_version__ as app_version$_ }
+export { app_version$_ as app_version__ }